refactor(domain): domain 로직 패키지 변경
기존에 구현한 클래스를 domain 패키지로 위치 변경
