+++ src/main/java/racingcar/domain/RacingStatus.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public enum RacingStatus {
     STOP,
